Moved precomputation of -A into EdDSAPublicKeySpec
For public keys that are used to verify a single message, this provides no
speed benefit, and requires the storage in memory of another GroupElement and
its dblPrecmp. For public keys that are used to verify more than one message,
the storage increase is the same, but the precomputation only occurs once now
instead of during every verification.