Use Ed25519ScalarOps instead of BigIntegerScalarOps
Before:
0.534 each sign, 2.3272 each verify, 2.8612 s+v
After:
0.5888 each sign, 2.3566 each verify, 2.9454 s+v

Sign uses reduce() twice, verify uses it once. Therefore the ref10 version is
about 28us slower than the BigInteger version. Basically negligible, so the
important factor to determine is whether the BigInteger operations used are
constant time. If we can't get the ref10 version to be constant time, we may as
well use the BigInteger version.