Implement one-shot methods in EdDSAEngine
The one-shot methods can be used to improve performance by not copying the data
if all the data is available. They are unsupported by the Java JCA and must be
used directly.

This commit also fixes an API violation if EdDSAEngine object is reused for
signing.

I2P ticket: https://trac.i2p2.de/ticket/1750

Signed-off-by: str4d <str4d@mail.i2p>